requirements.txt
google-generativeai>=0.5.0
openai>=1.0.0
pydantic>=2.0.0
jsonschema>=4.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
flask>=3.0.0
